Rename build configuration module for consistency
